refactor: rename relay notifier provider to relay provider, use timer.isActive instead of new bool field, move restartable timer to services
