+++ lib/app/features/nostr/providers/relays.c.dart
@@ -2,20 +2,19 @@
 
 import 'dart:async';
 
-import 'package:ion/app/utils/restartable_timer.dart';
+import 'package:ion/app/services/timer/restartable_timer.dart';
 import 'package:nostr_dart/nostr_dart.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
 
-part 'relays_provider.c.g.dart';
+part 'relays.c.g.dart';
 
 typedef RelaysState = Map<String, NostrRelay>;
 
 @Riverpod(keepAlive: true)
-class RelayNotifier extends _$RelayNotifier {
+class Relay extends _$Relay {
   late RestartableTimer _timer;
   late StreamSubscription<RelayMessage> _messagesSubscription;
   late StreamSubscription<int> _subscriptionsCountSubscription;
-  bool _timerFinished = false;
   int? _subscribersLength;
 
   @override
@@ -26,7 +25,7 @@ class RelayNotifier extends _$RelayNotifier {
     // 1. Has no active subscriptions
     // 2. it has been at least 30s since last message from the relay was received
     _initializeTimer(relay);
-    _messagesSubscription = relay.messages.listen((_) => _resetTimer());
+    _messagesSubscription = relay.messages.listen((_) => _timer.reset());
     _subscriptionsCountSubscription = relay.subscriptionsCountStream.listen((length) {
       _subscribersLength = length;
       _processUpdate(relay);
@@ -38,19 +37,11 @@ class RelayNotifier extends _$RelayNotifier {
   }
 
   void _initializeTimer(NostrRelay relay) {
-    _timer = RestartableTimer(const Duration(seconds: 30), () {
-      _timerFinished = true;
-      _processUpdate(relay);
-    });
-  }
-
-  void _resetTimer() {
-    _timerFinished = false;
-    _timer.reset();
+    _timer = RestartableTimer(const Duration(seconds: 30), () => _processUpdate(relay));
   }
 
   void _processUpdate(NostrRelay relay) {
-    if (_subscribersLength == 0 && _timerFinished) {
+    if (_subscribersLength == 0 && !_timer.isActive) {
       _timer.cancel();
       relay.close();
       ref.invalidateSelf();
