+++ pubspec.yaml
@@ -66,7 +66,7 @@ dependencies:
   nostr_dart:
     git:
       url: https://github.com/ice-blockchain/nostr-dart.git
-      ref: 0.0.11
+      ref: 0.0.12
   ogp_data_extract: ^0.1.4
   package_info_plus: ^4.2.0
   path: ^1.9.0
